requirements.txt
youtube_transcript_api 
langchain-community
google-generativeai 
langchain-google-genai
python-dotenv
unstructured 
load_dotenv
langchain
chromadb
tiktoken 
einops
flask
PyPDF
fpdf

